% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_interaction_subset.R
\name{bs_interaction_subset}
\alias{bs_interaction_subset}
\title{B-Spline Basis for the interaction series' subset that actually varies.}
\usage{
bs_interaction_subset(indic_interaction, x, ...)
}
\arguments{
\item{indic_interaction}{Boolean vector indicating which interaction series values are varying.}

\item{x}{Numeric vector of full interaction series (a subset exhibits meaningful variation).}

\item{...}{Arguments to \code{splines::bs}.}
}
\value{
A \code{splines::bs} object (operator).
}
\description{
Drop interaction series values which reflect no dynamics (switch turned off),
then compute B-Spline Basis for the result.
}
