% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns_order.R
\name{check_columns_order}
\alias{check_columns_order}
\title{Record/apply a data frame's columns order.}
\usage{
check_columns_order(df, filename_operator, run_type)
}
\arguments{
\item{df}{A data frame.}

\item{filename_operator}{Character filename for operator save.}

\item{run_type}{"train" or "test", dictating whether reference data saves to disk.}
}
\value{
A data frame, with ordered columns.
}
\description{
Record or apply an operator to enforce specific column names and order,
according to a 'train' reference.
}
