% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_categories_other.R
\name{consolidate_categories_other}
\alias{consolidate_categories_other}
\title{Consolidate categorical variable's rarely observed levels into 'other'.}
\usage{
consolidate_categories_other(
  df,
  varname_cat,
  threshold,
  value_consol = "OTHER",
  construct_indic_other = FALSE
)
}
\arguments{
\item{df}{A data frame, with categorical variable whose rarely observed levels should be consolidated.}

\item{varname_cat}{Character name of categorical variable.}

\item{threshold}{Minimum share of observations with categorical level \emph{v}, below which, \emph{v} enters 'other'.}

\item{value_consol}{Value which constitutes the newly consolidated 'other' category.}

\item{construct_indic_other}{Does construct binary indicator variable strictly to convey, 'is_other'.}
}
\value{
Modified version of input data frame.
}
\description{
Two options, controlled by \code{construct_indic_other}:
\itemize{
\item (FALSE) Construct modified version of original categorical variable.
Overwrite some level values with 'other', according to minimum frequency threshold.
Modified variable result takes '_consol' suffix.
\item (TRUE) Construct binary indicator variable strictly to convey, 'is other',
according to minimum frequency threshold.
}
}
