% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_sparse_other_cat.R
\name{consolidate_sparse_other_cat}
\alias{consolidate_sparse_other_cat}
\title{Consolidate sparsely-observed categorical levels, into 'other'.}
\usage{
consolidate_sparse_other_cat(
  df,
  varname_cat,
  threshold,
  value_consol = "OTHER",
  construct_indic_other = TRUE
)
}
\arguments{
\item{df}{Data frame with categorical variable whose sparsely-observed levels should be consolidated.}

\item{varname_cat}{Categorical variable name.}

\item{threshold}{Minimum share of observations with categorical variable value \emph{v}, if \emph{v} to be kept.}

\item{value_consol}{Categorical variable \emph{level} into which sparsely-observed values will be consolidated.}

\item{construct_indic_other}{Does construct an indicator variable strictly to convey, 'is_other'.}
}
\value{
Original data frame, with new consolidated categorical variable ('_consol' suffix).
}
\description{
Consolidate sparsely-observed categorical levels, into 'other'.
}
