% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_bspline_basis.R
\name{construct_bspline_basis}
\alias{construct_bspline_basis}
\title{Construct B-Spline basis functions.}
\usage{
construct_bspline_basis(x, knots, varnames_out_prefix)
}
\arguments{
\item{x}{Original predictor variable.}

\item{varnames_interact}{Vector of character variable names.}
}
\value{
B-spline ('bs') object.
}
\description{
Adopt HRW's ZOSull spline knots defaults,
and apply to splines' B-spline basis generator.
}
