% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_factor_interactions.R
\name{construct_factor_interactions}
\alias{construct_factor_interactions}
\title{Interact a factor-type variable with another variable}
\usage{
construct_factor_interactions(df, pair_varnames)
}
\arguments{
\item{df}{A data frame.}

\item{pair_varnames}{Character vector of variable names to be interacted.}
}
\value{
A data frame, containing the newly constructed interactions.
}
\description{
A factor-type variable interacted with another yields,
one new variable per factor level. New variable naming
follows a consistent \verb{factor level}:\code{variable} naming convention.
}
