% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_interact_char.R
\name{construct_interact_char}
\alias{construct_interact_char}
\title{Construct interaction of character variables, by pasting together.}
\usage{
construct_interact_char(df, varnames_interact)
}
\arguments{
\item{df}{Data frame with character variables to be interacted.}

\item{varnames_interact}{Vector of character variable names.}
}
\value{
Original data frame, with new interacted (pasted) variable.
}
\description{
Construct interaction of character variables, by pasting together.
}
