% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_vars_mdy.R
\name{construct_vars_mdy}
\alias{construct_vars_mdy}
\title{Construct separate month-day-year variables, from single combined date variable.}
\usage{
construct_vars_mdy(df, varnames_date)
}
\arguments{
\item{df}{Data frame with date variable(s) to be manipulated.}

\item{varnames_date}{Vector of original combined date variable names.}
}
\value{
Data frame with both original and new date variables.
}
\description{
Construct separate month-day-year variables, from single combined date variable.
}
