% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_factor_order_freq.R
\name{declare_factor_order_freq}
\alias{declare_factor_order_freq}
\title{Declare factor-type variable using underlying frequencies.}
\usage{
declare_factor_order_freq(df, varname_x)
}
\arguments{
\item{df}{A data frame with variable to become factor-type.}

\item{varname_x}{Character variable name.}
}
\value{
Modified version of input data frame.
}
\description{
Declare factor-type variable using underlying frequencies.
}
