% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_consolidate_categories_operator.R
\name{derive_consolidate_categories_operator}
\alias{derive_consolidate_categories_operator}
\title{Derive a categorical variable's operator for 'other' value consolidation.}
\usage{
derive_consolidate_categories_operator(
  df,
  varname_cat,
  thresh_nobs_consol,
  thresh_nobs_indic_other,
  value_consol = "OTHER",
  directory_operator
)
}
\arguments{
\item{df}{A data frame -- effectively, a 'train' reference.}

\item{varname_cat}{Character name of categorical variable.}

\item{thresh_nobs_consol}{For a modified version of original categorical variable --
minimum count of observations with categorical level \emph{v}, below which, \emph{v} enters 'other'.}

\item{thresh_nobs_indic_other}{For a binary indicator of 'other' --
minimum count of observations with categorical level \emph{v}, below which, \emph{v} enters 'other'.}

\item{value_consol}{For a modified version of original categorical variable -- v
value which labels the newly consolidated 'other' category.}

\item{directory_operator}{Character directory for operator save.}
}
\value{
A list, containing:
* A character filename of the constructed operator.
* A data frame, showing whether a category level consolidates to 'other'.
}
\description{
Execute any combination of 'consolidate_categories_other' executions (use NULL threshold to not run a particular type).
After determining which category levels consolidate into 'other', save those mappings as an operator.
}
