% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_standardization_operator.R
\name{derive_standardize_operator}
\alias{derive_standardize_operator}
\title{Derive a data frame standardization operator.}
\usage{
derive_standardize_operator(df, filename_operator)
}
\arguments{
\item{df}{A data frame.}

\item{filename_operator}{Character filename for operator save.}
}
\value{
A list, containing:
* The character filename of the operator.
* The operator object.
}
\description{
Given a 'train' reference with variables to be standardized (mean zero, unit-variance),
derive and save the transforming operator.
}
