% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_nzv.R
\name{drop_nzv}
\alias{drop_nzv}
\title{Drop features with near/exactly zero variance, aligned with 'train' reference.}
\usage{
drop_nzv(df, filename_operator, run_type)
}
\arguments{
\item{df}{A data frame.}

\item{filename_operator}{A complete file path for 'train' reference.}

\item{run_type}{"train" or "test", dictating whether reference data saves to disk.}
}
\value{
A data frame, whose columns are a subset of input columns.
}
\description{
Drop features with near/exactly zero variance, aligned with 'train' reference.
}
