% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_data_types.R
\name{examine_data_types}
\alias{examine_data_types}
\title{Examine data types, compare to 'train' reference.}
\usage{
examine_data_types(df, filename_operator, run_type)
}
\arguments{
\item{df}{A data frame.}

\item{filename_operator}{A complete file path for 'train' reference.}

\item{run_type}{"train" or "test", dictating whether reference data saves to disk.}
}
\value{
Vector: character entries reflect data types, names reflect variable names.
}
\description{
Collect data types in a data.frame.
If 'train' execution, save that summary.
If 'test' execution, confirm equality with 'train' reference.
}
