% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories_rare.R
\name{get_categories_rare}
\alias{get_categories_rare}
\title{Retrieve a categorical variable's rare levels.}
\usage{
get_categories_rare(df, varname_cat, thresh_frac_obs)
}
\arguments{
\item{df}{A data frame.}

\item{varname_cat}{A character name of categorical variable.}

\item{thresh_frac_obs}{Numeric fraction of observations below which, category level considered 'rare'.}
}
\value{
Character indicator variable names.
}
\description{
Retrieve a categorical variable's rare levels.
}
