% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interact_components.R
\name{get_interact_components}
\alias{get_interact_components}
\title{For an interaction series following from a binary indicator and second variable, extract component series.}
\usage{
get_interact_components(varname_interact, df)
}
\arguments{
\item{varname_interact}{Character variable name of the interaction series.}

\item{df}{A data frame, containing the interaction series and underlying on/off switch.}
}
\value{
A list, containing:
* The interaction series
* The on/off interaction indicator series
}
\description{
For an interaction series following from a binary indicator and second variable, extract component series.
}
