% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varnames_some_na.R
\name{get_varnames_some_na}
\alias{get_varnames_some_na}
\title{Get the names of data frame variables containing NA values.}
\usage{
get_varnames_some_na(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
Character vector.
}
\description{
Get the names of data frame variables containing NA values.
}
