% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_bar_simple.R
\name{ggplot_bar_simple}
\alias{ggplot_bar_simple}
\title{Quickly visualize one discrete variable's level frequencies.}
\usage{
ggplot_bar_simple(df, varname_x, order_x = TRUE, filename_print = NULL)
}
\arguments{
\item{df}{A data frame with plotting data.}

\item{varname_x}{Discrete variable name (character string).}

\item{order_x}{Do order variable levels along x-axis by frequency.}

\item{filename_print}{Character name of a plot file for save.}
}
\value{

}
\description{
Quickly visualize one discrete variable's level frequencies.
}
