% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_bar_x.R
\name{ggplot_bar_x}
\alias{ggplot_bar_x}
\title{ggplot a bar chart, with by-group flexibility and optional print-to-pdf.}
\usage{
ggplot_bar_x(
  df,
  varname_x,
  varname_grp = NULL,
  order_x = T,
  units_share = F,
  path_filename_print = NULL
)
}
\arguments{
\item{df}{Data frame with plotting data.}

\item{varname_x}{Variable name whose distribution will be plotted.}

\item{varname_grp}{Variable name used for grouping.}

\item{order_x}{Variable values ordered along axis by frequency.}

\item{units_share}{Vertical axis units are in terms of within-x value share.}

\item{path_filename_print}{Optional pdf print file}
}
\value{
ggplot plot object.
}
\description{
ggplot a bar chart, with by-group flexibility and optional print-to-pdf.
}
