% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_density_simple.R
\name{ggplot_density_simple}
\alias{ggplot_density_simple}
\title{Quickly visualize one continuous variable's distribution of values.}
\usage{
ggplot_density_simple(df, varname_x, filename_print)
}
\arguments{
\item{df}{Data frame with plot data.}

\item{varname_x}{Character variable name.}

\item{filename_print}{Character name of a plot file for save.}
}
\value{
A ggplot object.
}
\description{
Quickly visualize one continuous variable's distribution of values.
}
