% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_group_mean.R
\name{impute_group_mean}
\alias{impute_group_mean}
\title{Imputation using by-group mean.}
\usage{
impute_group_mean(x, varnames_impute, varnames_grp, pre_df_means = NULL)
}
\arguments{
\item{x}{Data frame with missing values, and those values for imputation calculations.}

\item{varnames_impute}{Vector of variable name(s) with missing values to be imputed.}

\item{varnames_grp}{Vector of variable name(s) used for grouping.}

\item{pre_df_means}{Existing data frame of by-group means, if available.}
}
\value{
A list of associated imputation objects:
\itemize{
\item (If pre_df_means not specified) df_means, data frame of by-group means.
\item M, data frame of T/F, indicating: this cell was missing originally.
\item df_impute, imputations overwritten for missing values of original input 'df'.
}
}
\description{
Imputation using by-group mean.
}
