% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_consolidate_categories_operator.R
\name{join_consolidate_categories_operator}
\alias{join_consolidate_categories_operator}
\title{Join a categorical variable's operator for 'other' value consolidation, with no-joins also declared 'other'.}
\usage{
join_consolidate_categories_operator(
  filename_operator,
  df,
  varname_cat,
  varname_suffix_consol = "consol",
  value_consol,
  varname_suffix_indic_other = "indic_other"
)
}
\arguments{
\item{filename_operator}{Filename of saved categorical variable consolidation operator.}

\item{df}{A data frame, with categorical variable to be consolidated.}

\item{varname_cat}{Character name of a categorical variable.}

\item{varname_suffix_consol}{Character suffix for categorical variable's consolidated version (contained in operator).}

\item{value_consol}{Value which constitutes the newly consolidated 'other' category.}

\item{varname_suffix_indic_other}{Character suffix for categorical variable's binary 'is other' version (contained in operator).}
}
\value{
Modified version of input data frame.
}
\description{
Categorical variable consolidation operator is a data frame, mapping some category levels to 'other'.
Read operator from disk, as portable summary of 'train' reference.
Given a new data frame with initial categorical variable,
join the operator to construct consolidated categorical variable.
}
