% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_filename_op_consol.R
\name{mk_filename_op_consol}
\alias{mk_filename_op_consol}
\title{Make consolidation operator filename, according to centralized naming convention.}
\usage{
mk_filename_op_consol(directory_operator, varname_cat)
}
\description{
Make consolidation operator filename, according to centralized naming convention.
}
