% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_posterior_beta_binomial.R
\name{plugin_posterior_beta_binomial}
\alias{plugin_posterior_beta_binomial}
\title{Plug-in update of beta prior under binomial likelihood.}
\usage{
plugin_posterior_beta_binomial(par_prior_beta, D)
}
\arguments{
\item{par_prior_beta}{List-type object, with elements "a" and "b" for beta prior parameters.}

\item{D}{Vector of 0-1 data (generated by binomial likelihood function).}
}
\value{
List with beta posterior parameters.
}
\description{
Plug-in update of beta prior under binomial likelihood.
}
