% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bs_interaction_subset.R
\name{predict_bs_interaction_subset}
\alias{predict_bs_interaction_subset}
\title{Evaluate Spline Basis for an interaction series subset that actually varies}
\usage{
predict_bs_interaction_subset(object, indic_interaction, newx)
}
\arguments{
\item{object}{A \code{splines::bs} object.}

\item{indic_interaction}{Boolean vector indicating which interaction series values are varying.}

\item{newx}{The interaction \code{x} values, containing a subset at which evaluations are required.}
}
\value{
A data frame of spline basis evaluations, saved within original interaction series' on/off form.
}
\description{
Drop interaction series values which reflect no dynamics (switch turned off),
then evaluate B-Spline Basis for the result. Save the evaluation
in the original interaction series' context, among values which reflect no dynamics.
}
