% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_plot_present.R
\name{relabel_plot_present}
\alias{relabel_plot_present}
\title{Polish chart labels for presentation, using variable names lookup table.}
\usage{
relabel_plot_present(
  p,
  varname_x0,
  df_relabels = NULL,
  title_prefix = NULL,
  lab_y = NULL,
  lab_x = "Value"
)
}
\arguments{
\item{p}{ggplot object.}

\item{varname_x0}{Initial variable name (to be polished).}

\item{df_relabels}{Lookup table matching original variable name to polished name.}

\item{title_prefix}{String for consistent chart titling.}

\item{lab_y}{String for presentable y-axis.}

\item{lab_x}{String for presentable x-axis.}
}
\value{
ggplot object.
}
\description{
Helpful primarily for scaled production of univariate charts.
}
