% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_univar_plot.R
\name{relabel_univar_plot}
\alias{relabel_univar_plot}
\title{Re-label univariate plot's noted variable name.}
\usage{
relabel_univar_plot(
  p,
  varname_x0,
  varnames_dict = NULL,
  title_prefix = NULL,
  lab_y = NULL,
  lab_x = "Value"
)
}
\arguments{
\item{p}{ggplot object.}

\item{varname_x0}{Initial (character string) variable name.}

\item{varnames_dict}{Dictionary (lookup table) matching original variable name (column 1) to polished name (column 2).}

\item{title_prefix}{Character for consistent chart titling.}

\item{lab_y}{Character for presentable y-axis.}

\item{lab_x}{Character for presentable x-axis.}
}
\value{
ggplot object.
}
\description{
Variable names not typically optimal for plot presentations.
Use reference dictionary for re-label to human-friendly language.
This function helps primarily when producing many univariate charts.
}
