% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot_fmt_pdf.R
\name{save_plot_fmt_pdf}
\alias{save_plot_fmt_pdf}
\title{Print pdf plot(s) formatted with often-helpful defaults.}
\usage{
save_plot_fmt_pdf(p, filename_print)
}
\arguments{
\item{p}{Plot object.}

\item{filename_print}{Character name of plot file for save.}
}
\value{
NULL.
}
\description{
Print pdf plot(s) formatted with often-helpful defaults.
}
