% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_store_Y.R
\name{separate_store_Y}
\alias{separate_store_Y}
\title{Separate Y from input dataset, save on disk.}
\usage{
separate_store_Y(df, varname_y, directory, run_type)
}
\arguments{
\item{df}{A data frame.}

\item{varname_y}{Character Y variable name.}

\item{directory}{Character folder name where Y saves.}

\item{run_type}{Character "train" or "test" (for filename).}
}
\value{
Modified input data frame, lacking Y variable.
}
\description{
Separate Y from input dataset, save on disk.
}
