% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df.R
\name{split_df}
\alias{split_df}
\title{Split a data frame and return either train or test partition.}
\usage{
split_df(df, fraction_train, run_type)
}
\arguments{
\item{df}{A data frame.}

\item{fraction_train}{Fraction of the dataset declared 'train'.}

\item{run_type}{Character indicating partition returned, 'train' or 'test'.}
}
\value{

}
\description{
Split a data frame and return either train or test partition.
}
