% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_summary_stats.R
\name{stack_summary_stats}
\alias{stack_summary_stats}
\title{Compute summary statistics and stack in an output data frame, with a "summary stat" label variable.}
\usage{
stack_summary_stats(
  df,
  varnames_group,
  varnames_summarize = "all",
  funs_summary,
  filename_operator
)
}
\arguments{
\item{df}{A data frame.}

\item{varnames_group}{A character vector of by-groups.}

\item{varnames_summarize}{A character vector of numeric variables to be summarized.}

\item{funs_summary}{List of numeric-geared functions, named.}

\item{filename_operator}{A complete file path for 'train' reference.}
}
\value{
Data frame of summary statistics.
}
\description{
Compute summary statistics and stack in an output data frame, with a "summary stat" label variable.
}
