% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_var_discrete.R
\name{tab_var_discrete}
\alias{tab_var_discrete}
\title{Tabulate a discrete variable.}
\usage{
tab_var_discrete(df, varname, rename_generic = FALSE)
}
\arguments{
\item{df}{Data frame with variable to be tabulated.}

\item{varname}{Variable name to be tabulated.}

\item{rename_generic}{In tabulation result, does output
a variable name-generic df, with columns: varname,
value, n.}
}
\value{
Data frame - a tabulation report.
}
\description{
With a plan to append variable tabulation result
with other variables' tabulation results,
force to character, avoiding clashing data types.
}
