% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_frequency_overall.R
\name{tabulate_frequency_overall}
\alias{tabulate_frequency_overall}
\title{Tabulate frequencies for grouping variable(s).}
\usage{
tabulate_frequency_overall(df, varnames_group)
}
\arguments{
\item{df}{A data frame.}

\item{varnames_group}{Character name(s) of variables for frequency tabulation.}
}
\value{
A data frame with summary frequencies.
}
\description{
Tabulate frequencies for grouping variable(s).
}
