% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_var_discrete.R
\name{tabulate_var_discrete}
\alias{tabulate_var_discrete}
\title{Tabulate discrete variable(s).}
\usage{
tabulate_var_discrete(df, var)
}
\arguments{
\item{df}{Data frame with variable to be tabulated.}

\item{var}{Character variable name(s) to be tabulated.}

\item{rename_generic}{In tabulation result, does output
a variable name-generic df, with columns: varname,
value, n.}
}
\value{
Data frame - a tabulation report.
}
\description{
With a plan to append variable tabulation result
with other variables' tabulation results,
force to character, avoiding clashing data types.
}
