% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_bar_x.R
\name{ggplot_bar_x}
\alias{ggplot_bar_x}
\title{ggplot a bar chart, with by-group flexibility and optional print-to-pdf.}
\usage{
ggplot_bar_x(
  df,
  varname_x,
  ...,
  order_x = TRUE,
  units_share = FALSE,
  path_filename_print = NULL
)
}
\arguments{
\item{df}{Data frame with plotting data.}

\item{varname_x}{Variable name (as character string) whose distribution will be plotted.}

\item{...}{Other named arguments -- character string input values -- passing to ggplot2's aesthetics.}

\item{order_x}{Variable values are ordered along axis by frequency.}

\item{units_share}{Vertical axis units are in terms of within-x value share.}

\item{path_filename_print}{Optional pdf print file}
}
\value{
ggplot plot object.
}
\description{
ggplot a bar chart, with by-group flexibility and optional print-to-pdf.
}
